+++ Changelog.md
@@ -1,5 +1,5 @@
 ## 0.0.63
-+ Added `muna.beta.openai.embeddings.create` method for creating embeddings with text embedding predictors using the OpenAI API.
++ Added `muna.beta.openai.embeddings.create` method for using text embedding predictors via the OpenAI API.
 + Added `Parameter.EmbeddingDims` annotation method to specify embedding Matryoshka dimension parameters.
 + Refactored `Parameter.modality` field to `Parameter.denotation`.
 + Renamed `muna --learn` CLI option to `muna --docs`.